Solving type error of y_data.

+++ generate_to_torch/splited_Pneumonia_all_true_false_split_1_(300, 300).py
@@ -1,4 +1,5 @@
 from utils import DataModule, NeuralNetwork, TrainModule
+from tensorflow.keras.utils import to_categorical
 from torch.utils.tensorboard import SummaryWriter
 from torchinfo import summary
 from torch import optim
@@ -28,7 +29,7 @@ for key in nploader:
 train_x_data = nploader["train_x_data"] / 255.0
 train_x_data = np.expand_dims(train_x_data, axis=1)
 print(np.shape(train_x_data))
-train_y_data = nploader["train_y_data"].astype(np.long)
+train_y_data = nploader["train_y_data"]
 train_loader = dm.np_to_dataloader(train_x_data, train_y_data)
 del train_x_data
 del train_y_data
@@ -36,7 +37,7 @@ del train_y_data
 valid_x_data = nploader["valid_x_data"] / 255.0
 valid_x_data = np.expand_dims(valid_x_data, axis=1)
 print(np.shape(valid_x_data))
-valid_y_data = nploader["valid_y_data"].astype(np.long)
+valid_y_data = nploader["valid_y_data"]
 valid_loader = dm.np_to_dataloader(valid_x_data, valid_y_data)
 del valid_x_data
 del valid_y_data
@@ -44,7 +45,7 @@ del valid_y_data
 test_x_data = nploader["test_x_data"] / 255.0
 test_x_data = np.expand_dims(test_x_data, axis=1)
 print(np.shape(test_x_data))
-test_y_data = nploader["test_y_data"].astype(np.long)
+test_y_data = nploader["test_y_data"]
 test_loader = dm.np_to_dataloader(test_x_data, test_y_data)
 del test_x_data
 del test_y_data
